patch provided by Michael D.:
I’ve bumped into a problem with Tasker2 and Daylight Saving Time. 

The symptom is that tasks keep getting rescheduled even though they have
already run.  I tracked the problem down to the local time not honoring DST,
so the local time comes back one hour earlier than actual.  This
particularly occurs with the combination of DST being in effect, a task
interval of 0100, and a task that takes less than 1 minute to complete.

The root of the problem is that under Windows Mobile, the _localtime64_s()
function does not honor DST.  This is true regardless of the setting of
HomeDST in the registry, or state of TIME_ZONE_INFORMATION.  The result is
that the local time reported by _localtime64_s() is one hour earlier than
actual, and the next scheduled task time is the current time instead of one
hour in the future.

...the call to nclog() on line 117 of myNotify.c in the wasctime_s() function
is missing an argument for the error code.

PATCH provided:
55a56,65
> 
>       // Adjust the local time with valid DST info since _localtime64_s doesn't honor HomeDST.
>       // Don't bother with year, minute or seconds, since DST can't change them.
>       // Don't bother with tm_yday because it isn't used in code.
>       SYSTEMTIME local_time;
>       GetLocalTime(&local_time);
>       pLocalTime->tm_hour = local_time.wHour;
>       pLocalTime->tm_mday = local_time.wDay;
>       pLocalTime->tm_wday = local_time.wDayOfWeek;
>